Added simple inventory class
Added things for testing purpose, doesn't affect gameplay.
Created a new class called inventory, added some simple logic to it. 
Modified gamelogic so you can display what is in your inventory.
Added to string in weapon class to allow it to be displayed through the inventory class
Did the same to the potion class